+++ CHANGELOG.md
@@ -2,9 +2,24 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+## Added
+- Ability to change font and font size of application. 
+  - This font change is saved to the config file. 
+- ability to change font in main window.
+
+## Fixed
+- changed "name" to "Name" in the main tracker table header.
+
+## Changed
+### Mac
+- Changed button style so they can resize. (Sorry rounded corner fans.)
+- Main window buttons resize on Mac
+
 ## [0.3.0] 2016-01-25
 ### Added
 - Custom Gambits can now be added from the Edit menu. These are added to the config file.
+-  "Custom Gambit" option now included in decisive window where you are allowed to enter your own difficulty.
 - Onslaught Penalty appears in the main window.
 
 ### Fixed